move config into page
